Update and rename lobbytransfer.js to lobbygive.js
Added some debugging.
